Transform settings to fragment for bottom bar
