+++ app/build.gradle
@@ -33,6 +33,7 @@ android {
 
 dependencies {
     implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.preference:preference:1.0.0'
 
     implementation 'pl.bclogic:pulsator4droid:1.0.3'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.3'
