+++ app/src/main/res/layout/activity_lesson_expandable_list_group.xml
@@ -10,7 +10,7 @@
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
         android:layout_margin="3dp"
-        android:fontFamily="@font/googlesans_medium"
+        android:fontFamily="@font/googlesans_regular"
         android:paddingStart="?android:attr/expandableListPreferredItemPaddingLeft"
         android:textSize="17sp" />
